feat(www:store): foreign key reference updates and docs
