+++ www/store/features/meeting.js
@@ -1,6 +1,6 @@
 import meetingAPI from "../../api/meeting";
-import { generateActions } from "../utils/slice-generator";
-import { generateSelectors } from "../utils/slice-generator";
+import { generateActions } from "../normalized-store/normalized-store";
+import { generateSelectors } from "../normalized-store/normalized-store";
 
 export const schema = {
   name: "meeting",
