+++ www/store/root.js
@@ -1,18 +1,19 @@
 import { combineReducers } from "redux";
 
-import { createNormalizedReducer } from "./utils/slice-generator";
+import { createReducer } from "./normalized-store/normalized-store";
 
 import userReducer from "./features/user";
 import themeReducer from "./features/theme";
 import drawerReducer from "./features/drawer";
 import snackbarReducer from "./features/snackbar";
+
 import topic from "./features/topic";
 import meeting from "./features/meeting";
 import takeaway from "./features/takeaway";
 import actionItem from "./features/action-item";
 import participant from "./features/participant";
 
-const normalizedReducer = createNormalizedReducer({
+const normalizedReducer = createReducer({
   topic: topic.schema,
   meeting: meeting.schema,
   takeaway: takeaway.schema,
